changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11771 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/DataExplorer/ORRecordIndexer.h
@@ -46,8 +46,8 @@
 - (BOOL) cancelAndStop;
 - (void) setFileToReplay:(NSString*)newFileToReplay;
 - (void) setDataRecords: (NSArray *) aDataRecords;
-- (void) setTotalLength:(unsigned)aLength;
-- (void) setLengthDecoded:(unsigned)aLength;
+- (void) setTotalLength:(NSUInteger)aLength;
+- (void) setLengthDecoded:(NSUInteger)aLength;
 - (void) setHeader:(id)aHeader;
 - (void) checkStatus;
